UI/WebStore/Controllers/WebApiController.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;
using Microsoft.AspNetCore.Mvc;
using WebStore.Interfaces.Api;

namespace WebStore.Controllers
{
    public class WebApiController : Controller
    {
        private readonly IValuesService _valuesService;

        public WebApiController(IValuesService valuesService)
        {
            _valuesService = valuesService;
        }

        public async Task<IActionResult> Index()
        {
            return View(await _valuesService.GetAsync());
        }
    }
}